+++ imagebuildah/chroot_symlink_unsupported.go
@@ -1,0 +1,7 @@
+// +build !linux
+
+package imagebuildah
+
+func resolveChrootedSymlinks() {}
+
+func resolveSymlinkTimeModified() {}
