+++ run_darwin.go
@@ -1,0 +1,5 @@
+package buildah
+
+func setChildProcess() error {
+	return ErrNotSupported
+}
